extract config to separate file
